Replace isinstance(.., str) with isinstance(.., six.string_types)
Checking isinstance(.., str) is deprecated in Python 3, in favor of isinstance(.., basestring). Since basestring is not a builtin in Python 2, a common workaround is using six.string_types. An alternative would be a try/catch block with a catch for the "NameError" exception (around basestring), but I think this is cleaner.